refactor(engine): resolve SonarCloud findings for the #1666 backend binding
Address the 11 SonarCloud MAJOR code smells the PR introduced, without changing
behavior:

- Keep the three touched files under the 500-line limit by extracting cohesive
  helpers into their own modules: the legacy ownership-evidence resolver
  (provisioner range_backend_evidence.py), the write-once binding helpers
  (engine services _range_backend_binding.py), and the live-fire admission gate
  (cms services _range_backend_admission.py, re-exported from _range_create).
- Reduce _resolve_operation_backend to <=3 returns; add the missing
  _uses_active_gdc_range_plane docstring; compare the write-once binding by
  scalar fields (clears the always-true dict-comparison finding).
- Suppress S6553 (# NOSONAR) on the two deliberate null=True sentinel columns
  (NULL distinguishes "no binding" from a value, per the preflight).
- Hoist non-throwing setup out of pytest.raises blocks in the new tests.

Part of #1666.
