+++ shifter/engine/provisioner/provisioner_db.py
@@ -460,58 +460,6 @@ def get_aces_range_data_by_request_id(request_id: str) -> dict[str, Any]:
     }
 
 
-# GDC VM Runtime asset types (VM Runtime guests + scenario pods) vs the GCE VM
-# range-cell asset type. This is the durable ownership discriminant persisted on
-# engine_instance.state; used only to resolve legacy (pre-#1666) ranges that have
-# no explicit backend binding (#1666).
-_GDC_ASSET_TYPES = frozenset({"vm_runtime_vm", "scenario_pod"})
-_GCE_ASSET_TYPE = "gce_vm"
-
-
-def resolve_legacy_range_backend(request_id: str) -> str | None:
-    """Resolve a legacy (NULL-binding) GCP range's backend from ownership evidence (#1666).
-
-    Inspects the durable ``asset_type`` discriminant persisted on the range's
-    ``engine_instance.state`` rows: ``gce_vm`` proves the GCE range-cell backend,
-    ``vm_runtime_vm``/``scenario_pod`` prove the GDC VM Runtime backend. Returns
-    the proven backend only when the evidence is unambiguous (exactly one backend
-    across all request-owned instances); returns ``None`` for an empty, mixed, or
-    unrecognized set so the caller fails closed (a ``prerequisite`` denial +
-    operator backfill) rather than guessing from the mutable env selector. Names,
-    scenario shape, current selector, and successful VM boot are NOT evidence.
-    """
-    with get_db_connection() as conn, conn.cursor() as cur:
-        cur.execute(
-            """
-            SELECT ei.state
-            FROM engine_instance ei
-            JOIN engine_request er ON ei.request_id = er.id
-            WHERE er.request_id = %s
-            """,
-            (request_id,),
-        )
-        rows = cur.fetchall()
-
-    backends: set[str] = set()
-    for (state,) in rows:
-        if isinstance(state, str):
-            try:
-                state = json.loads(state)
-            except (TypeError, ValueError):
-                continue
-        if not isinstance(state, dict):
-            continue
-        asset_type = str(state.get("asset_type", "")).strip()
-        if asset_type == _GCE_ASSET_TYPE:
-            backends.add("gce")
-        elif asset_type in _GDC_ASSET_TYPES:
-            backends.add("gdc")
-
-    if len(backends) == 1:
-        return next(iter(backends))
-    return None
-
-
 def get_aces_image_candidates(provider: str, source_name: str) -> list[dict[str, Any]]:
     """Return enabled ACES image mappings for (provider, source_name) (ADR-032-R2).
